Modify to take fuzzy ranking format form algorithm

+++ api/internal/domain/VendorRanking.go
@@ -1,8 +1,8 @@
 package domain
 
 type VendorRanking struct {
-	ProjectID int `json:"projectId"`
-	VendorID  int `json:"vendorId"`
-	Ranking   int `json:"ranking"`
-	Score     int `json:"score"`
+	ProjectID int     `json:"projectId"`
+	VendorID  int     `json:"vendorId"`
+	Ranking   int     `json:"ranking"`
+	Score     float64 `json:"score"`
 }
